move decorators out of WidgetBase
