+++ src/decorators/afterRender.ts
@@ -1,0 +1,14 @@
+import { handleDecorator } from './handleDecorator';
+
+/**
+ * Decorator that can be used to register a function to run as an aspect to `render`
+ */
+export function afterRender(method: Function): (target: any) => void;
+export function afterRender(): (target: any, propertyKey: string) => void;
+export function afterRender(method?: Function) {
+	return handleDecorator((target, propertyKey) => {
+		target.addDecorator('afterRender', propertyKey ? target[propertyKey] : method);
+	});
+}
+
+export default afterRender;
