+++ src/decorators/beforeRender.ts
@@ -1,0 +1,14 @@
+import { handleDecorator } from './handleDecorator';
+
+/**
+ * Decorator that can be used to register a reducer function to run as an aspect before to `render`
+ */
+export function beforeRender(method: Function): (target: any) => void;
+export function beforeRender(): (target: any, propertyKey: string) => void;
+export function beforeRender(method?: Function) {
+	return handleDecorator((target, propertyKey) => {
+		target.addDecorator('beforeRender', propertyKey ? target[propertyKey] : method);
+	});
+}
+
+export default beforeRender;
