+++ tests/unit/decorators/all.ts
@@ -1,2 +1,5 @@
+import './afterRender';
 import './beforeProperties';
+import './beforeRender';
+import './diffProperty';
 import './inject';
